Moved useResizeObserver to utils (#867)
