+++ packages/@react-spectrum/utils/src/index.ts
@@ -19,3 +19,5 @@ export * from './useDOMRef';
 export * from './styleProps';
 export * from './Slots';
 export * from './useHasChild';
+export * from './useResizeObserver';
+
