+++ packages/@spectrum-icons/color/tsconfig.types.json
@@ -11,7 +11,8 @@
   },
   "include": [
     "src",
-    "../../@react-types/shared/src/css.d.ts"
+    "../../@react-types/shared/src/css.d.ts",
+    "../../@react-spectrum/utils/src/ResizeObserver.d.ts"
   ],
   "exclude": ["scripts", "node_modules", "stories", "chromatic"]
 }
